# Performance improvements to events, stats and service_checks