This commit reduces allocations across the code by writing directly to the Buffer stream instead of creating intermediate vars. It also safeguards g.Tags from being modified by append calls. This is done by using a temp array sized according to the number of tags set. Also added a benchmark test for service_checks and events.

As an example the events allocation(/op) has dropped from 25 to 3.

New:-
```
13:19:41 DeepakMacbook: > go test -check.b -check.bmem
PASS: events_test.go:189: TestSuite.BenchmarkEvents	  500000	      5566 ns/op	     304 B/op	       3 allocs/op
PASS: stats_test.go:223: TestSuite.BenchmarkIncr	  500000	      5504 ns/op	     336 B/op	       7 allocs/op
OK: 2 passed
PASS
ok  	github.com/redsift/hilt/vendor/github.com/PagerDuty/godspeed	5.788s
```

Old:-
```
13:22:45 DeepakMacbook: > go test -check.b -check.bmem
PASS: events_test.go:189: TestSuite.BenchmarkEvents	  200000	     11714 ns/op	   12962 B/op	      25 allocs/op
PASS: stats_test.go:223: TestSuite.BenchmarkIncr	  200000	      6693 ns/op	     577 B/op	       8 allocs/op
OK: 2 passed
PASS
ok  	github.com/PagerDuty/godspeed	4.009s
```

Note: This could be further optimised by not creating additional intermediate arrays (it can be directly written to stream) and avoiding strings.Join (these can be directly written to stream as well).