@theckman 2018-01-17
@deepakprabhakara as long as I can do it without affecting your authorship of the commit, that's no problem. I'm pretty sure it'll keep you as the author and me as the committer, which is 👍 by me.

I hope to be back at my workstation in a few hours so that I can give this one final once-over and then merge. Thank you again.