@theckman 2018-01-18
@akrylysov Thank you so much for taking time to look at this PR too, and no worries! I found it the hard way (local tests exploded)... 😉 

* https://github.com/theckman/godspeed/pull/5

Thank you again.

@deepakprabhakara I'll be cutting `v1.1.0` soon, which will include these changes. I found some further optimizations I can make (will happen in PR #6), and I'll cut `v1.1.0` after I get a chance to merge them in.